Changed blockchainId from type Int to type Biginteger to accomodate uint256 from the agreement

+++ Smart-Contract-Verifier-Server/src/main/kotlin/com/savannasolutions/SmartContractVerifierServer/negotiation/models/Agreements.kt
@@ -3,6 +3,7 @@ package com.savannasolutions.SmartContractVerifierServer.negotiation.models
 import com.savannasolutions.SmartContractVerifierServer.messenger.models.Messages
 import com.savannasolutions.SmartContractVerifierServer.contracts.models.Judges
 import com.savannasolutions.SmartContractVerifierServer.user.models.User
+import java.math.BigInteger
 import java.util.*
 import javax.persistence.*
 
@@ -10,7 +11,7 @@ import javax.persistence.*
 data class Agreements(@Id @GeneratedValue val ContractID:UUID,
                       var AgreementTitle: String = "Un-named Agreement",
                       var AgreementDescription: String = "",
-                      var blockchainID: Int? = null,
+                      var blockchainID: BigInteger? = null,
                       val CreatedDate:Date,
                       var SealedDate:Date? = null,
                       var DurationConditionUUID: UUID? = null,
